Fixed an issue wih horpro casting

+++ utilityFunctions/preProcessing.py
@@ -3,7 +3,6 @@ import numpy as np
 import skimage.transform as trans
 from scipy.signal import argrelextrema
 from skimage.graph import route_through_array
-from skimage.viewer import ImageViewer
 from sklearn.neighbors import KernelDensity
 
 
@@ -535,7 +534,8 @@ def wordSegmentation(image, lineBreaks, maximas):
         horPro = np.sum(line, 0)
 
         # Most frequent value (MFV), which represents the width of the baseline
-        MFV = np.bincount(horPro[horPro != 0]).astype(np.uint8).argmax()
+        horPro = horPro.astype('int64')
+        MFV = np.bincount(horPro[horPro != 0]).argmax()
 
         # Hanshouf fein amaken el zeros
         # El mafroud ba3d el smoothing yetla3li amaken el zeros
